requirements.txt
# Django framework
Django>=5.2.4

# HTTP requests
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0

# Image processing
Pillow>=10.0.0

# HTML parser for BeautifulSoup
lxml>=4.9.0

# URL parsing utilities
urllib3>=2.0.0

# Production server
gunicorn>=21.2.0

# Environment variables
python-decouple>=3.8

# Database (PostgreSQL for production)
psycopg2-binary>=2.9.0

# Static files handling
whitenoise>=6.6.0
